# Save memory on histograms: slightly faster with less code