* one allocation for duo instantiations
* use index overflow instead of redundant last state field

name                         old time/op    new time/op    delta
HistogramWithLabelValues-12     204ns ± 0%     204ns ± 0%   ~     (all equal)
HistogramNoLabels-12           41.1ns ± 0%    40.5ns ± 0%   ~     (p=1.000 n=1+1)
HistogramObserve1-12           33.1ns ± 0%    33.9ns ± 0%   ~     (p=1.000 n=1+1)
HistogramObserve2-12            132ns ± 0%     122ns ± 0%   ~     (p=1.000 n=1+1)
HistogramObserve4-12            377ns ± 0%     364ns ± 0%   ~     (p=1.000 n=1+1)
HistogramObserve8-12           1.00µs ± 0%    0.96µs ± 0%   ~     (p=1.000 n=1+1)
HistogramWrite1-12             1.80µs ± 0%    1.76µs ± 0%   ~     (p=1.000 n=1+1)
HistogramWrite2-12             3.28µs ± 0%    3.25µs ± 0%   ~     (p=1.000 n=1+1)
HistogramWrite4-12             6.66µs ± 0%    6.66µs ± 0%   ~     (p=1.000 n=1+1)
HistogramWrite8-12             13.7µs ± 0%    13.6µs ± 0%   ~     (p=1.000 n=1+1)

name                         old alloc/op   new alloc/op   delta
HistogramWithLabelValues-12     48.0B ± 0%     48.0B ± 0%   ~     (all equal)
HistogramNoLabels-12            0.00B          0.00B        ~     (all equal)

name                         old allocs/op  new allocs/op  delta
HistogramWithLabelValues-12      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
HistogramNoLabels-12             0.00           0.00        ~     (all equal)