@beorn7 2019-02-10
Note: The test failure is unrelated to this PR. It's caused by a Travis change, which I'll take care of separately.