@pascaldekloe 2019-02-11
I've borrowed some of your thinking so this is fair-trade. 😁